+++ index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/public/images/icon/menu/feel-cat.svg" />
+    <link rel="icon" type="image/svg+xml" href="images/icon/menu/feel-cat.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>하루한냥</title>
   </head>
