+++ src/lib/utils/getValidationUser.ts
@@ -1,0 +1,15 @@
+const getValidationUser = (name: 'email' | 'password' | 'name', value: string) => {
+  if (name === 'email' || name === 'password' || name === 'name') {
+    const regexp = {
+      email: /^[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])*@[0-9a-zA-Z]([-_.]?[0-9a-zA-Z])*.[a-zA-Z]{2,3}$/i,
+      password: /^.{8,}$/,
+      name: /^.{2,}$/,
+    };
+
+    return regexp[name].test(value);
+  }
+
+  return false;
+};
+
+export default getValidationUser;
