+++ src/routes.test.tsx
@@ -28,11 +28,19 @@ describe('App', () => {
     });
   });
 
-  context('When the current page is "/signup"', () => {
-    it('renders the signup page', () => {
-      renderRouter(PATH.SIGN_UP);
+  // context('When the current page is "/signup"', () => {
+  //   it('renders the signup page', () => {
+  //     renderRouter(PATH.SIGN_UP);
+  //
+  //     screen.getAllByText(/회원가입/);
+  //   });
+  // });
 
-      screen.getAllByText(/회원가입/);
+  context('When the current page is "/calendar"', () => {
+    it('renders the calendar page', () => {
+      renderRouter(PATH.CALENDAR);
+
+      screen.getAllByText(/calendar/);
     });
   });
 });
